YieldMax/requirements.txt
python-dotenv==1.0.0
requests==2.31.0
pandas==2.0.0
numpy==1.24.0
pytest==7.0.0
